requirements.txt
matplotlib==3.8.4
Levenshtein
NumPy
Pillow
openai